jasper-colin/fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "motor==3.3.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.4.2",
    "pydantic-settings>=2.2.1",
    "pymongo>=4.11.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "uvicorn==0.24.0",
]
